+++ packages/yoga/src/Tag/Tag.theme.js
@@ -1,21 +1,24 @@
-const Tag = ({ spacing, fontSizes, fontWeights, radii }) => ({
+const Tag = ({ spacing, fontSizes, fontWeights, radii, borders }) => ({
   icon: {
+    size: spacing.small,
     margin: {
       right: spacing.xxxsmall,
     },
   },
   font: {
     size: fontSizes.xsmall,
-    weight: fontWeights.bold,
+    lineHeight: fontSizes.medium,
+    weight: fontWeights.medium,
   },
   padding: {
-    top: spacing.xxsmall,
+    top: spacing.xxxsmall,
     right: spacing.xxsmall,
-    bottom: spacing.xxsmall,
+    bottom: spacing.xxxsmall,
     left: spacing.xxsmall,
   },
   border: {
-    radius: radii.rounded,
+    width: borders.small,
+    radius: radii.semiRounded,
   },
 });