+++ packages/yoga/src/Tag/index.native.js
@@ -1,3 +1,6 @@
-import Tag from './native';
+import Tag from './native/Tag';
+import TagInformative from './native/Informative';
+
+Tag.Informative = TagInformative;
 
 export default Tag;