+++ packages/yoga/src/Tag/native/Tag.jsx
@@ -1,116 +1,75 @@
 import React from 'react';
-import styled, { withTheme } from 'styled-components';
-import { func, oneOf, oneOfType, bool, node } from 'prop-types';
-import { hexToRgb } from '@gympass/yoga-common';
-import { TriangleAlert } from '@gympass/yoga-icons';
+import styled from 'styled-components';
+import { oneOf, node } from 'prop-types';
+import Text from '../../Text';
 
-import { Text } from '../..';
+export const StyledTag = styled.View`
+  justify-content: center;
+  align-items: center;
 
-const StyledTag = styled.View`
   ${({
-    color,
-    full,
+    variant,
     theme: {
       yoga: {
+        colors: {
+          elements,
+          feedback: {
+            [variant]: borderColor = { dark: elements.selectionAndIcons },
+          },
+        },
         components: { tag },
       },
     },
   }) => `
-    width: ${full ? '100%' : 'auto'};
+    width: auto;
     padding:
       ${tag.padding.top}px
       ${tag.padding.right}px
       ${tag.padding.bottom}px
       ${tag.padding.left}px;
 
-    background-color: ${hexToRgb(color, 0.25)};
     border-radius: ${tag.border.radius}px;
+    border-width: ${tag.border.width}px;
+    border-color: ${borderColor.dark};
   `}
 `;
 
-const Wrapper = styled.View`
-  flex-direction: row;
-  justify-content: center;
-  align-items: center;
-`;
-
-const StyledText = styled(Text.Bold)`
+export const StyledText = styled(({ variant, ...rest }) => <Text {...rest} />)`
   ${({
-    color,
+    variant,
     theme: {
       yoga: {
+        colors: {
+          text,
+          feedback: { [variant]: color = { dark: text.secondary } },
+        },
         components: { tag },
       },
     },
   }) => `
-    color: ${color};
+    color: ${color.dark};
 
     font-size: ${tag.font.size}px;
+    line-height: ${tag.font.lineHeight}px;
+    font-weight: ${tag.font.weight};
   `}
 `;
 
 /** Use Tag component when you want to categorize your content */
-const Tag = ({
-  icon: Icon,
-  children,
-  variant,
-  theme: {
-    yoga: {
-      components: { tag },
-      colors: { text, [variant]: color = text.secondary },
-    },
-  },
-  ...props
-}) => (
-  <StyledTag color={color} {...props}>
-    <Wrapper>
-      {Icon && (
-        <Icon
-          width={14}
-          height={12}
-          fill={color}
-          style={{ marginRight: tag.icon.margin.right + 1 }}
-        />
-      )}
-
-      <StyledText color={color}>{children}</StyledText>
-    </Wrapper>
+const Tag = ({ children, variant, ...props }) => (
+  <StyledTag variant={variant} {...props}>
+    <StyledText variant={variant}>{children}</StyledText>
   </StyledTag>
 );
 
 Tag.propTypes = {
-  icon: oneOfType([bool, func]),
-  full: bool,
-  /** style the card following the theme (primary, secondary, vibin, hope,
-   * energy, relax, peace, verve, uplift, deepPurple, deep, stamina, dark,
-   * medium, light, clear, white) */
-  variant: oneOf([
-    '',
-    'primary',
-    'secondary',
-    'vibin',
-    'hope',
-    'energy',
-    'relax',
-    'peace',
-    'verve',
-    'uplift',
-    'deepPurple',
-    'stamina',
-    'dark',
-    'medium',
-    'deep',
-    'light',
-    'clear',
-    'white',
-  ]),
+  /** style the tag following the theme (success, informative, attention) */
+  variant: oneOf(['', 'success', 'informative', 'attention']),
   children: node.isRequired,
 };
 
 Tag.defaultProps = {
-  icon: TriangleAlert,
-  full: false,
   variant: '',
 };
 
-export default withTheme(Tag);
+export default Tag;