+++ packages/yoga/src/Tag/native/Tag.test.jsx
@@ -15,20 +15,24 @@ describe('<Tag />', () => {
     expect(container).toMatchSnapshot();
   });
 
-  it('should match snapshot with custom icon', () => {
+  it('should match snapshot with custom icon and informative type', () => {
     const { container } = render(
       <ThemeProvider>
-        <Tag icon={Building}>default with custom icon</Tag>
+        <Tag.Informative variant="success" icon={Building}>
+          success with custom icon
+        </Tag.Informative>
       </ThemeProvider>,
     );
 
     expect(container).toMatchSnapshot();
   });
 
-  it('should match snapshot with without icon', () => {
+  it('should match snapshot with without icon and informative type', () => {
     const { container } = render(
       <ThemeProvider>
-        <Tag icon={false}>default without icon</Tag>
+        <Tag.Informative variant="success">
+          success without icon
+        </Tag.Informative>
       </ThemeProvider>,
     );
 
@@ -38,33 +42,21 @@ describe('<Tag />', () => {
   it('should match snapshot with variant prop', () => {
     const { container } = render(
       <ThemeProvider>
-        <Tag variant="primary">primary</Tag>
-        <Tag variant="secondary">secondary</Tag>
-        <Tag variant="vibin">vibin</Tag>
-        <Tag variant="hope">hope</Tag>
-        <Tag variant="energy">energy</Tag>
-        <Tag variant="relax">relax</Tag>
-        <Tag variant="peace">peace</Tag>
-        <Tag variant="verve">verve</Tag>
-        <Tag variant="uplift">uplift</Tag>
-        <Tag variant="deepPurple">deepPurple</Tag>
-        <Tag variant="stamina">stamina</Tag>
-        <Tag variant="dark">dark</Tag>
-        <Tag variant="medium">medium</Tag>
-        <Tag variant="deep">deep</Tag>
-        <Tag variant="light">light</Tag>
-        <Tag variant="clear">clear</Tag>
-        <Tag variant="white">white</Tag>
+        <Tag variant="success">success</Tag>
+        <Tag variant="informative">informative</Tag>
+        <Tag variant="attention">attention</Tag>
       </ThemeProvider>,
     );
 
     expect(container).toMatchSnapshot();
   });
 
-  it('should match snapshot with full prop', () => {
+  it('should match snapshot with variant prop and informative type', () => {
     const { container } = render(
       <ThemeProvider>
-        <Tag full>full</Tag>
+        <Tag.Informative variant="success">success</Tag.Informative>
+        <Tag.Informative variant="informative">informative</Tag.Informative>
+        <Tag.Informative variant="attention">attention</Tag.Informative>
       </ThemeProvider>,
     );