+++ packages/yoga/src/Text/native/Text.jsx
@@ -43,6 +43,12 @@ Bold.displayName = 'Text.Bold';
 const Black = styledText('black');
 Black.displayName = 'Text.Black';
 
+const SectionTitle = styledText('sectionTitle');
+SectionTitle.displayName = 'Text.SectionTitle';
+
+const SmallestException = styledText('smallestException');
+SmallestException.displayName = 'Text.SmallestException';
+
 const TextRenderer = styledText('p');
 
 const Text = props => <TextRenderer {...props} />;
@@ -106,4 +112,6 @@ export {
   Medium,
   Bold,
   Black,
+  SectionTitle,
+  SmallestException,
 };